src/requirements/common.txt
Django==1.7.4
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
celery==3.1.17
coverage==3.7.1
django-bower==5.0.2
django-braces==1.4.0
django-extensions==1.5.0
django-nose==1.3
django-pipeline==1.4.5
django-selenium==0.9.8
django.js==0.8.1
futures=2.2.0
gevent==1.0.1
greenlet==0.4.5
gunicorn==19.2.1
j2cli==0.3.1-0
Jinja2==2.7.3
kombu==3.0.24
MarkupSafe==0.23
meld3==1.0.0
nose==1.3.4
psycopg2==2.5.4
python-gnupg==0.3.7
pytz==2014.10
selenium==2.44.0
six==1.9.0
sqlparse==0.1.14
supervisor==3.1.3
unicodecsv==0.9.4
wsgiref==0.1.2
